+++ lib/is-signed-in.js
