+++ lib/sign-in.js
