+++ lib/sign-out.js
