+++ lib/sign-up.js
